pyproject.toml
[project]
name = "dash-template"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash-mantine-components>=0.15.2",
    "dash>=2.18.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true

